+++ teste.py
@@ -23,89 +23,29 @@ kuma.plot(prop)
 
 kuma.metrics(prop)
 #%%
-def AD_beta(x, a, b):
-        cdf =  stats.beta.cdf(np.sort(x), a = a, b = b)
-        y = stats.norm.ppf(cdf)
-        u = stats.norm.cdf((y-np.mean(y))/np.std(y))
-        # aux = (np.concatenate(np.fromiter(((2*np.where(np.isclose(u, element))[0] -1)*np.log(element) + 
-        #                                     (2*len(x) - 2*np.where(np.isclose(u, element))[0] +1)*np.log(1-element) 
-        #                                     for element in u), object)))
-        aux = np.zeros(len(y))
-        i = 0
-        for u_ in u:
-            aux[i] = (2*i-1)*np.log(u_) + (2*len(x) - 2*i+1)*np.log(1-u_)
-            i = i + 1
-        a2 = -len(x) -(1/len(x))*sum(aux) 
-        ad = a2*(1+0.75/len(x) + 2.25/(len(x)**2))
-        return ad
 
-def AD_norm(x, loc, scale):
-        cdf = stats.norm.cdf(np.sort(x), loc = loc, scale = scale)
-        y = stats.norm.ppf(cdf)
-        u = stats.norm.cdf((y-np.mean(y))/np.std(y))
-        # aux = (np.concatenate(np.fromiter(((2*np.where(np.isclose(u, element))[0] -1)*np.log(element) + 
-        #                                     (2*len(x) - 2*np.where(np.isclose(u, element))[0] +1)*np.log(1-element) 
-        #                                     for element in u), object)))
-        aux = np.zeros(len(y))
-        i = 0
-        for u_ in u:
-            aux[i] = (2*i-1)*np.log(u_) + (2*len(x) - 2*i+1)*np.log(1-u_)
-            i = i + 1
-        a2 = -len(x) -(1/len(x))*sum(aux) 
-        ad = a2*(1+0.75/len(x) + 2.25/(len(x)**2))
-        return ad
-
-def AIC(likelihood, num_parameters):
-    aic = 2 * likelihood + 2 * num_parameters
-    return aic
-
-def BIC(likelihood, num_parameters, length):
-    bic = 2 * likelihood + num_parameters * np.log(length)
-    return bic
-
-def CAIC(likelihood, num_parameters, length):
-    aicc = 2 * likelihood + 2 * num_parameters * length / (
-        length - num_parameters - 1
-    )
-    return aicc
-
-#%%
-def llnorm(par, data):
-    mu, sigma = par
-    ll = -np.sum(np.log(2*np.pi*(sigma**2))/2 + ((data-mu)**2)/(2 * (sigma**2)))
-    return ll
-
-def betaNLL(par, data):
-    """
-    Negative log likelihood function for beta
-    <param>: list for parameters to be fitted.
-    <args>: 1-element array containing the sample data.
-
-    Return <nll>: negative log-likelihood to be minimized.
-    """
-
-    a, b = par
-    pdf = stats.beta.pdf(data,a,b,loc=0,scale=1)
-    lg = np.log(pdf)
-    mask = np.isfinite(lg)
-    nll = -lg[mask].sum()
-    return nll
 #%%
 
 mle = minimize(llnorm, x0=theta0, method="Nelder-Mead", args=(prop))
-stats.cramervonmises(prop, 'beta',  args=(mle.x[0], mle.x[1]))
-stats.kstest(prop, 'beta',  args=(mle.x[0], mle.x[1]))
+res = stats.cramervonmises(prop, 'norm',  args=(mle.x[0], mle.x[1]))
+res1 = stats.kstest(prop, 'norm',  args=(mle.x[0], mle.x[1]))
 AD_norm(prop, mle.x[0], mle.x[1])
 AIC(mle.fun, len(mle.x))
 BIC(mle.fun, len(mle.x), len(prop))
 CAIC(mle.fun, len(mle.x), len(prop))
+criterios = kuma.metrics(prop)
+criterios_kuma = [criterios['AIC'], criterios['AICc'], criterios['BIC'], criterios['cramer_vonmises'], criterios['AD'], criterios['KS']]
+criterios_normal = ([AIC(mle.fun, len(mle.x)), BIC(mle.fun, len(mle.x), len(prop)), CAIC(mle.fun, len(mle.x), len(prop)),
+                    AD_norm(prop, mle.x[0], mle.x[1]), res.statistic, res1.statistic])
 #%%
 mle = minimize(betaNLL, x0=theta0, method="Nelder-Mead", args=(prop))
-stats.cramervonmises(prop, 'beta',  args=(mle.x[0], mle.x[1]))
-stats.kstest(prop, 'norm',  args=(mle.x[0], mle.x[1]))
+res = stats.cramervonmises(prop, 'beta',  args=(mle.x[0], mle.x[1]))
+res1 = stats.kstest(prop, 'beta',  args=(mle.x[0], mle.x[1]))
 AD_beta(prop, mle.x[0], mle.x[1])
 AIC(mle.fun, len(mle.x))
 BIC(mle.fun, len(mle.x), len(prop))
 CAIC(mle.fun, len(mle.x), len(prop))
+criterios_beta = ([AIC(mle.fun, len(mle.x)), BIC(mle.fun, len(mle.x), len(prop)), CAIC(mle.fun, len(mle.x), len(prop)),
+                    AD_beta(prop, mle.x[0], mle.x[1]), res.statistic, res1.statistic])
 # stats.cramervonmises(prop, 'beta')
 #%%
